Add files via upload
An real estate website using HTML,CSS and PHP where user can register their account using a username and password and also login to their account . All the images of the real estate property are present with a description and also location using google map. User can also view and animation in the homepage with all the uses of HTML and CSS.
Place all the files in your server and its ready to work.